Factorise Map logic
